requirements.txt
openai>=0.28.0
python-dotenv>=1.0.0
fastapi>=0.103.1
uvicorn>=0.23.2
requests>=2.31.0
beautifulsoup4==4.12.2
PyPDF2==3.0.1
python-multipart>=0.0.6
pillow==10.1.0
httpx>=0.25.2
jinja2==3.1.2
pydantic>=2.3.0
starlette>=0.27.0
cachetools>=5.3.1
numpy>=1.24.0
chromadb>=0.4.18 